(edit) rename and update the component 'Input' to 'TextInput' and adjust its strory
